requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.0.6
exceptiongroup==1.2.1
executing==2.0.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
ipython==8.24.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
requests==2.32.2
rsa==4.9
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
